+++ src/main/java/io/github/imagineDevit/giwt/TestCaseCtxResult.java
@@ -13,26 +13,58 @@ import java.util.Objects;
  */
 public class TestCaseCtxResult<T> extends ATestCaseResult<T> {
 
+    /**
+     * Private constructor that accepts a value
+     *
+     * @param value the value to be wrapped in the TestCaseCtxResult
+     */
     private TestCaseCtxResult(T value) {
         super(value);
     }
 
+    /**
+     * Private constructor that accepts an Exception
+     *
+     * @param e the Exception to be wrapped in the TestCaseCtxResult
+     */
     private TestCaseCtxResult(Exception e) {
         super(e);
     }
 
+    /**
+     * Factory method to create a new TestCaseCtxResult with a given value
+     *
+     * @param value the value to be wrapped in the TestCaseCtxResult
+     * @return a new TestCaseCtxResult instance with the given value
+     */
     protected static <T> TestCaseCtxResult<T> of(T value) {
         return new TestCaseCtxResult<>(value);
     }
 
+    /**
+     * Factory method to create a new TestCaseCtxResult with a null value
+     *
+     * @return a new TestCaseCtxResult instance with a null value
+     */
     protected static <T> TestCaseCtxResult<T> empty() {
         return new TestCaseCtxResult<>(null);
     }
 
+    /**
+     * Factory method to create a new TestCaseCtxResult with a given Exception
+     *
+     * @param e the Exception to be wrapped in the TestCaseCtxResult
+     * @return a new TestCaseCtxResult instance with the given Exception
+     */
     protected static <T> TestCaseCtxResult<T> ofErr(Exception e) {
         return new TestCaseCtxResult<>(e);
     }
 
+    /**
+     * Method to get the result of the TestCaseCtxResult
+     *
+     * @return a TestCaseResult instance with the value or Exception of this TestCaseCtxResult
+     */
     protected TestCaseResult<T> result() {
         return Objects.requireNonNull(value, "Result value is Null")
                 .<T>ok()
